vdk-core: Add flag to JobConfig in case config file is required (#2521)
Currently, the vdk-core's JobConfig does not require the presense of a
config file in a data job's repository. This works in general, but not
for the vdk-control-cli's use-cases, where the presense of the config
file is mandatory.

This change adds a `should_fail_missing_config` flag to JobConfig, which
can be used when the presense of the config file is necessary. If the
flag is set to `True`, and the config file is missing in the data job's
folder, a VdkConfigurationError will be raised.

Testing Done: Added test

Signed-off-by: Andon Andonov <andonova@vmware.com>