vdk-core: add vdk sql-query command (#2512)
Add vdk sql-query command which will use the same database as the rest
of the SQL interfaces (SQL steps, job_input methods)

This will replace the per-db commends like vdk impala-query, vdk
sqlite-query which are just problematic.

This would enable to generalize a lot of the tutorials/examples we have
which currently give impression they can be used only wiht "sqlite"
(sqlite-query) or trino. Which is primary driver for doing the change
now.

The command would be used in functional tests where you want to query
the default db without needing ot create a job for that.

The command for now is hidden (hidden feature) since there are a few
TODOs that would be resolved in subsequent PRs.