vdk-audit: [bug fix] Fix incorrectly detected event (#2548)
The logic in the `_audit()` method filters the forbidden events and
detects if an executed event is allowed or not. However, the way it
matches the events is by partial string match (the `event in
not_permitted_event` part). This means that if a forbidden event is,
e.g., `os.removexattr` and the event that is passed to the method is
`os.remove`, there will be a match, because `os.remove` is a sub-string
of `os.removexattr`.

With this change, the `_audit()` method switches to exact string match
when filtering for forbidden events, in order to avoid situations where
unexpected events are matched.

Testing done: Added test

Signed-off-by: Andon Andonov <andonova@vmware.com>