vdk-jupyter: add oauth2 authentication implementation
This is adding the server part of Oauth2 authentication process.

It adds 1 more APIs: `/login`

When called it without "code" query paramter, it will start the
authentication proces as per OAuth2 standard .
We are using only native app workflow with PKCE (RFC 7636) because we
cannot really secure the server side so we cannot reliably use client
secret.
When called with "code" query paramter it will finish the process and
exchange the code for access token (and refresh token) and safe it in
VDK storage.

This change add integration with jupyter configuration. This way the
extension can be configured more natively using jupyter configuration
mechanism.

In future change we can add integration between VDK configuration
mechanims and jupyter so that properties set in VDK can be recognized in
Jupyter and vice-versa but that's more advanced use-case