vdk-jupyter: Change root FS position for path parameter
Currently, VDK commands in Jupyter require a path parameter which
points to the location of the data job relatively to the current
location of the user in the Jupyter file system menu on the right.

This is suboptimal, as it makes it less consistent to use and is the
cause of at least one bug.

This change streamlines this by making it so the path parameter is
always assumed to point relatively to the FS location where Jupyter
was instantiated. This means all commands will expect the same path
param value for a given job, regardless of where the user is currently
located within the Jupyter FS menu.

This also fixes a bug where a job inside a given directory cannot be
ran from inside that directory.

Testing done: only manual so far, working on an actual test

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>