vdk-audit: Set tryfirst on plugin configuration hook
Currently, there are random cases, where the `vdk_configure` hook in the audit plugin is evaluated before
the configuration hook in custom plugins that set the same configuration variables.

This change sets the *tryfirst* option of the hook to `True` as an attempt to force the hook to be
evaluated before custom plugins' configuration hooks. We do this, as there is currently no mechanism
to specify the order of hook evaluations.

Testing Done: N/A

Signed-off-by: Andon Andonov <andonova@vmware.com>