vdk-control-cli: Add python_version to sample config.ini
This change adds the `python_version` property to the config.ini file of the sample data
job that is returned when a user executes `vdk create -n <job-name> -t <team-name>`. The
property is commented out, as it is used only if a user wants to set a python version
different from the default one.

Testing Done: Executed `vdk create -n my-test-job -t my-team` locally, and verified that
the python_version property is present in the generated data job's config.ini file.

Signed-off-by: Andon Andonov <andonova@vmware.com>