control-service: remove hardcoded image pull policy from job deployer (#2557)
what: removed hardcoded imagePullPolicy from the JobImageDeployer
service.

why: 
Currently, the "Always" strategy fetches the latest version of the image
from the registry. This results in an unnecessary image pulls.
Switching to the "IfNotPresent" strategy will address this issue. With
this strategy, images will be pulled only if they are not already
present in the local cache, matching the behavior of our data jobs that
use exact tags associated with the current version. This will reduce
unnecessary network calls to the Docker Registry.

testing: trivial

---------

Signed-off-by: mrMoZ1 <mzhivkov@vmware.com>