frontend: change cypress version in test docker image to 9.7.0
Why

https://github.com/vmware/versatile-data-kit/pull/2358 bumped the
version of cypress to 12. This requires test migration, however,
so we have to revert it for now.

Also changing back the angular-cli version to de-risk something
similar happening with angular.

What

Change cypress version back to 9.7.0
Change angular-cli-version back to 13.3.7

How was this tested

CI

What kind of change is this

Bugfix

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>