vdk-impala: Introduce COMPUTE STATS statements (#2584)
Why:
We were facing issues for large sized tables sporadically when using the
processing templates with specified quality checks.
The issues appear when we try to move the data from the staging table
(where the quality checks were already performed) into the target table.
The absence of "COMPUTE STATS" statement against the staging table was
causing failures when trying to select data from it due to hitting query
limits. Before submitting this PR here are the checks that have been
done in order to take that decision:
-A job that was processing a large table with quality checks defined was
ran 45 times from which only 17 were successful and other were failures
due to the issue mentioned above. After that a "Compute stats" statement
was added against the staging table before trying to process the data to
target - the result was 100 consecutive successful runs. That's why we
assumed this enhancement will optimize the execution of the templates.

More details explained in
https://github.com/vmware/versatile-data-kit/issues/1361

What:
-Adding COMPUTE STATS statement that will be executed against the
staging table right before moving the data to prod in order to optimize
further selects on it.

Signed-off-by: Stefan Buldeev
[sbuldeev@vmware.com](mailto:sbuldeev@vmware.com)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>