vdk-core: pass vdk run arguments in standalone job mode as well (#2601)
When running StandaloneDataJob we do not pass `vdk run` arguments to
`vdk_start` hook, but some hook implementations rely on that to
understand what command is being run without that it breaks their logic.

JobConfigIni Configuraiton provider is one such hook implementation See

https://github.com/vmware/versatile-data-kit/blob/main/projects/vdk-core/src/vdk/internal/builtin_plugins/config/vdk_config.py#L189

As can be seen we infer the job paht from the command line and try to
find config.ini there.
This is hacky (there's unfixed TODO) and naturally causes such issues as
hacky things do. But we have not come up with better way then that.

This is causing issues with Jupyter integration where config.ini file is
not being parsed and read!

To fix that and for consistency since StandaloneDataJob simulates vdk
run of a job we pass all the arguments to vdk_start hook as if vdk run
command was run.