vdk-control-api-auth: vdk credentials cache refactoring
This is refactoring the credentials cache storage system to make it more
reusable . Particulary I want to reuse it to get corrected authenticated
user in telemetry later.

- We are providing explcit interface for credentials cache intead of
using Union of classes
- Implement SingletonInMemoryCredentials cache to enable different parts
of the program to easily access and reuse the same credentials.
- Added method read_cached_access_token_only and use it in
get_authenticated_username() to make it easier to record user using vdk
even if his token has expired.
- Default credentials_cache to LocalFolderCredentialsCache which is more
commonly used.
- Deprecate cache_locally in favour of explicitly passing the credential
cache instance, this is better design pattern.