vdk-jupyter: improve init message
It's better not to provide link to git repo source code. Links to source
code are not best documentation. But also job_input could be changed
through plugins in which case it won't be accurate.

Notebook `help(xxx)` is much more suitable.

Remove the unnecesary warning. it's not critical and even if by mistake
they mark the current cell , they'd quickly find out t problem when
using vdk run