vdk-core: make sure standalone data job doesn't run steps (#2609)
Standalone Data Job ignores directory provided steps and doesn't run
them (at least not currently).

But if a plugin adds more steps , this logic does not work and it runs
the steps adde by the plugin. In our case vdk-notebook add new notebook
steps and they are run . This causes when a new notebook is created and
job_input is intiailzied , during initializaiton it tries to run all
steps in the notebook.

This fixes this issue by adding explicit function after initialize_job
that replaces all steps with noop step.