vdk-core: adopt plugin 1.3
Plugy realease of 1.3 on 25th of August introduced a changing renaming
pluggy._callers._Result to pluggy.Result.

It can be found in the list of changes between 1.2 and 1.3 :
https://github.com/pytest-dev/pluggy/compare/1.2.0...1.3.0

As we rely on that as a type for the Result object , it breaks vdk.

This changing is adopting the pluggy change