VDK-DuckDB: Introducing a new database plugin (#2561)
<VDK-plugin for the DuckDB Database.>


# Why?
This plugin has been created to check the differences how databases are
implemented in a VDK plugin.
The aim of this change is to add DuckDB to the list of database plugins.

# What?
I made a new DuckDB-plugin based on the SQL-plugin.
For making the DuckDB work I had to change several things.
For the Plugin I had to change: Module Imports, Configuration,
Connection Factory Method, Ingester Method and Library Specific Code.
For the Ingestion I had to change: Import Statements, Configuration and
connection, Target Path, Data Type Conversion, Query compatibility,
General naming and references, Connection Handling.
For the Configuration I had to change: Constants, Class Names, Method
for Database File, Configuration Key, Descriptions.
For the Connection I had to change: Classes, Class documentation, File
Name, Import Statement, Connection Method, Isolation Level.

# How has this been tested?
I've made a test program that first relied on the 'auto create feature'
of the plugin. But after several errors I realized
that probably the auto create function wasn't functioning as expected. 
So now the test doesn't rely on this function and instead doesn't create
the table in advance, disables auto create, checks if the table is
missing
and makes sure the ingestion logic remains.

# What type of change are you making?
This is a new feature.
Now the users can make use of the DuckDB database from the list of
database plugins.

---------

Co-authored-by: Maximiliaan <maxsobry@gmai.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>