versatile-data-kit: add poster image to all the videos
I am adding a "poster image" or placeholder image to all the videos (which really is the first frame of the video so simply edit each video through canva to add first 0.1 seconds with still image. 

I also edited a bit the getting started and ingestion videos while I was doing that.