vdk-jupyter: improve init message (#2618)
It's better not to provide link to git repo source code. Links to source
code are not best documentation. But also job_input could be changed
through plugins in which case it won't be accurate.

Notebook `help(xxx)` is much more suitable.

Remove the unnecesary warning. it's not critical and even if by mistake
they mark the current cell , they'd quickly find out t problem when
using vdk run

I also made the cell non-editable. During user testing, users were
trying to edit it and this may cause potential issues as it's not
expected

output of help(job_input) looks like that: 


![image](https://github.com/vmware/versatile-data-kit/assets/2536458/5edac2d8-c32b-469e-9ed2-84cb4c72827c)