vdk-jupyter: show status on starting new operation
Common usability issue with starting operations is that when you click
the operation (e.g Run) nothing happens.

This is fairly cheap temporary fix to at least show the status dialog
and ask the user to explicitly close it.

Longer term solution should be planned where we will not only show that
operaiton is runnign, but show how long time it reemians, where it is
currently (e.g what is in case of vdk run), show logs (if user wants to)
and so on. But that's a bigger story.