control-service: add data job deployment dynamic property source selection mechanism (#2641)
what:
Added a feature flag mechanism which will be used to allow for writes to
both K8S and the database, and providing the flexibility to select the
data source (either the database or K8S) when reading properties based
on the feature flag configuration.

why: 
This change anticipates future work on:
https://github.com/vmware/versatile-data-kit/tree/main/specs/vep-2272-complete-data-job-configuration-persistence#database-data-model-changes.
which aims to remove the dependency on the Kubernetes (K8S) cluster for
storing part of the configuration of data jobs. By decoupling the
configuration from the K8S cluster, we can improve the backup/restore
process and overall system performance. Currently, the loss of K8S
namespace results in the potential loss of data jobs, requiring a manual
and complex restoration process.

testing:
added unit test which checks default configuration

---------

Signed-off-by: mrMoZ1 <mzhivkov@vmware.com>
Co-authored-by: github-actions <unknown>
Co-authored-by: Andy <87700128+doks5@users.noreply.github.com>
Co-authored-by: github-actions <>