vdk-core: test  ingest with multiple methods
To address a lack of teseting for  where payloads with different
"methods" were
may be improperly aggregated together.

The way it works this happen by creating distinct IngesterBase instances
for each method . But this was not realy tested so we could not be sure
it works. And also the logic is a bit strange. We pass method to
IngesterBase , but then we actually ignore it

As a small improvement, introduced a `PayloadObject` data class to
encapsulate payload-related variables, enhancing code readability and
maintainability.