vdk-heartbeat: Add python version configuration
Currently, it is not possible to set a specific python version, with which a test data job
to be deployed. As a result the heartbeat can be executed only against the default python
version specified in the Control Service configuration.

This change adds the ability to set a specific python version in the ini file for a test case.

Testing Done: Executed vdk-heartbeat with a simple test config.ini and verified that the test
data job was deployed with thhe correct python version.

Signed-off-by: Andon Andonov <andonova@vmware.com>