vdk-jupyter: improve init message (#2688)
It seems the instructions to run the first cell are being missed often.

So I've created a markdown cell where we can put in bold the
instructions so they are more noticeable.

Beyound that since we always "edit" a new notebook if hte plugin is
installed. It's good idea to explain why so if a user create a notebook
but doesn't know the plugin is installed , they understand why do they
have all those extra things inside.

Also added an empty cell so user can get stared. Without an empty cell
and , there are only 2 non-editable cells and if the user doesn't know
how to create a new cell or just tries to edit the non-ediable cells, he
would be very confused what to do.


Screenshot: 

![image](https://github.com/vmware/versatile-data-kit/assets/2536458/c1c0db6a-7d47-4185-bf80-09b0803211cf)