vdk-jupyter: Enable PYTHONUNBUFFERED to ensure correct log ordering
Currently, logs made inside jobs using print statements are not
positioned correctly within the vdk_logs file, but rather print
logs are appended at the end of the file.

This change fixes this by enabling the PYTHONUNBUFFERED env
variable to ensure that print statements are not buffered
and are written to the file immediately.

Testing done: local test

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>