vdk-notebook: set summary file path as configuration (#2709)
Previously, the summary file path was hard-coded to the job directory,
posing issues for certain deployments where writing to the job directory
was restricted. This led to failures in some deployments.

The second problem. This design introduced a hidden and unsafe
dependency between the vdk-notebook and vdk-jupyter plugins. The
vdk-notebook plugin was responsible for writing the summary information,
while vdk-jupyter read from the same hard-coded location, although both
plugins operated independently.

This commit migrates the summary file path logic to a more appropriate
location.The functionality to output summary of the job in a way that
differnet consumers (plugins or tools) can read and understand is rather
generic and core functionality regardless if vdk-notebook is installed
or not.

After merging this would change vdk-jupyter to work with this and remove
the obsolete code in vdk-noteboook