control-service: add new deployment tables (#2719)
# Why
As part of the VEP-2272, we have to introduce a process to synchronize
the data jobs from the database to the Kubernetes.

The initial implementation depends on the data_job_deployment table for
both reading and writing operations. As the deployment process operates
asynchronously with unpredictable durations, users may believe that
their deployment has been completed while it may not have.
To address this issue, we'll be introducing a new table called
"desired_data_job_deployment" and renaming the current one to
"actual_data_job_deployment". The read operations will read from
"actual_data_job_deployment" while the write operations will write to
"desired_data_job_deployment".

# What
We've implemented new deployment tables.

Just so you know, this is just the second phase of implementation. The
following enhancements are planned for future PRs:

- Once the new tables have been utilized, the old one will be deleted.
- We will annotate the method DataJobsSynchronizer.synchronizeDataJobs()
with @scheduled.
- Improved exception handling will be integrated.
- More tests will be included in subsequent updates.
- ThreadPool configuration will be tuned and exposed through the
application.properties.

Testing Done
Integration tests

Signed-off-by: Miroslav Ivanov
[miroslavi@vmware.com](mailto:miroslavi@vmware.com)

---------

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com
Signed-off-by: Miroslav Ivanov [miroslavi@vmware.com](mailto:miroslavi@vmware.com)
Co-authored-by: github-actions <>