vdk-core: pass execution result through context
Why?

We'd like to have just one exit point for data jobs where error
logging happens and propagate all errors up to it.

What?

Pass execution result in job context when exceptions are thrown.
This way, we still get to throw the exception and have the result
at a later hook, e.g. we have the execution result from run_job
inside the run method, even if an exception was thrown during
the run. If an exception was thrown, we can catch and log it at
the exit point.

The side effect is that plugins have to check the job context
for hook implementations that are wrapping if an exception was
thrown.

How was this tested?

Ran tests locally
Ran plugin tests in CI/CD

What kind of change is this?

Potentially breaking change