vdk-control-cli: pin werkzeug to version 2.3.8 or less
Why?

A test is failing due to changes in the werkzeug API.

https://gitlab.com/vmware-analytics/versatile-data-kit/-/jobs/5203253935

What?

Pin the werkzeug version to 2.3.8 or less, where the
API we use in the test is still supported.

How was this tested?

Ran test locally and confirmed that it passes

What kind of change is this

Bugfix

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>