vdk-impala: add Out Of Memory error handling (#2747)
Add Out Of Memory error handling logic in the vdk-impala error handler.

The Impala query planner projects the memory usage for every statement based on existing statistics about the
tables involved in the statement. This statistics could be either unavailable or incorrect which leads to wrong
memory estimates. We try to tackle this problem by gradually increasing the per executor node memory limit for
the query and retrying it. As a last resort, we set the memory limit very high. Neither of these actions should
cause problems for a well configured Admission Control in an Impala instance.

For more information:
https://impala.apache.org/docs/build/html/topics/impala_admission.html
https://impala.apache.org/docs/build3x/html/topics/impala_mem_limit.html