control-service: data job synchronizer error handling (#2742)
Why
As part of VEP-2272, we need to introduce a process for synchronizing
data jobs from the database to Kubernetes. In the event of a data job
deployment failure (due to user or platform errors), the current
implementation attempts to deploy the data job during every
synchronization cycle.

What
We have added a data job deployment status field to the
desired_data_job_deployment table. This status is used to determine
whether the deployment failed in the previous cycle and should be
skipped in the current one. The failed deployment status can be reset by
invocation of updateDeployment API event without a job code change (it
will be implemented as part of
https://github.com/vmware/versatile-data-kit/pull/2731/files).

This is just the third phase of implementation. The following
enhancements are planned for future PRs:

- We will annotate the method DataJobsSynchronizer.synchronizeDataJobs()
with @scheduled.
- Improved exception handling will be integrated.
- More tests will be included in subsequent updates.
- ThreadPool configuration will be tuned and exposed through the
application.properties.

Testing done
Unit tests

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com

---------

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com
Co-authored-by: github-actions <>