vdk-core: add error formatter configuration
Why?

Different exception message formats are required for different use
cases, e.g. CLI vs. Jupyter vs. Cloud

What?

Add a flag to vdk-core configuration that lets the user pick the
formatter. Initial choices are "pretty" and "plain". The plain
format should be good enough for Jupyter and Cloud deployments.
The pretty formatter is enabled by default, because CLI needs
to work out of the box.

Formatters have been exported to fuctions in errors.py, which
makes it easy to write a new formatter if necessary. Exceptions
are logged only at the exit point in the `run` method, which is
also where formatting takes place.

How was this tested?

Caused some config and user errors in local runs and looked at
the output

CI tests

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>