vdk-core: handle stack traces when passing custom components to internal code
In https://github.com/vmware/versatile-data-kit/issues/2620
the described use scenario is when user using
`send_tabular_data_for_ingeston(rows=BrokenIterator())`

The user provided broken iterator is called within plaform code and the
issue is is it recongized as user error.

It is correctly. There's a test already covering that scenario
[test_run_user_error_fail_job_ingest_iterator](https://github.com/vmware/versatile-data-kit/blob/634febe37cab75ffc36cd88ef186b7c910b8e03c/projects/vdk-core/tests/functional/run/test_run_errors.py#L77)

The test ErrorClassifierTest are not needed because the logic that
classifies that type of error as user error is not done by stack trace
but by handling directly the exception.