vdk-core: simplify error message for send_**_for_ingestion
There were a couple of issues with error handling that are fixedhere

1. Inconsistent behaviour between send_xxx methods. One method was
swallowing the error while the other was re-throwing it . This is fixe
by both rethrowing it.

2. Remove the unecessary error message. It is redundant and doesn't
provide any new information that is not availbe in the raised exception

3. When "rows" argument was not iterabot nor cursor or iterable raised
custom exceptin,  it always failed with cryptic error ` AttributeError:
'BrokenIterator' object has no attribute 'fetchmany' `
This is fixed by using more appropriate if statement to cehck type of
rows and raising InvalidArgumentsIngestionException if it's the wrong
type.

Consolited some tests in function/ingestion module.
There was no test covering the scenario of using multiple methods and it
is added.