control-service: fix control service post deployment test (#2790)
# Why
We observed the following post-deployment test failure:
https://gitlab.com/vmware-analytics/versatile-data-kit/-/jobs/5192553226.

This change is the culprit
https://github.com/vmware/versatile-data-kit/pull/2714
It’s the first pipeline that started having this issue
https://gitlab.com/vmware-analytics/versatile-data-kit/-/jobs/5210024385
The previous change for the control service that triggered the same
tests passed fine. There weren’t any changes to heartbeat or control-cli
in between, as far as I can tell. You can see from the logs that the
schedule is written to the data job table for the previous control
service change that triggered the same tests.


https://gitlab.com/vmware-analytics/versatile-data-kit/-/pipelines/1020826286
https://gitlab.com/vmware-analytics/versatile-data-kit/-/jobs/5192553226

The control-cli does a PUT request after creating the data job. This
updates the config, which includes the schedule. For some reason, the
config doesn’t get updated. The change touches some of the update code,
namely, it passes the new DesiredJobDeployment entity into the
JobImageBuilder, which is used in the JobService update method, which
gets called when we do a PUT request to update the job.

# What
Removed the unnecessary @Transactional annotation from the
DesiredJobDeploymentRepository methods which prevent from committing the
transaction started in JobsService.updateJob().

# Testing done
Manually created data job through the DataJob API and then updated it
through the DataJob API.

create: localhost:8092/data-jobs/for-team/taurus/jobs
```json
{
	"job_name": "miroslavi-test6",
	"team": "taurus",
        "config": {
          "schedule": {
	          "schedule_cron": "* * * * *"
          }
	}
}
```

update: localhost:8092/data-jobs/for-team/taurus/jobs/miroslavi-test6
```json
{
	"team": "taurus",
	"config": {
          "schedule": {
	          "schedule_cron": "* * * * 1"
          }
	}
}
```

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com

---------

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com
Co-authored-by: github-actions <>