vdk-structlog: create structured logging plugin
Why?

As part of VEP-2448, we'd like to make structured logging pluggable

What?

Create a vdk-structlog plugin. The plugin allows users to configure
logging metadata and logging format. It also works with bound loggers.

How was this tested?

Installed the plugin and ran a data job locally with
different configurations

White kind of change is this?

feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>