control-service: add integration tests for async job deployment
I've introduced the initial version of async job deployment integration tests. This is just the starting point, and we plan to expand and refactor it further in the upcoming version based on feedback from aivanov's comment. You can find the discussion on this change in the following link: Link to the GitHub pull request discussion

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com