control-service: asynchronous deployment deletion (#2781)
Why:
As part of VEP-2272, we need to introduce a process for synchronizing
data jobs from the database to Kubernetes. Currently, the process lacks
the ability to delete deployments asynchronously.

What:
We have introduced an asynchronous method for deployment deletion. It is
based on the logic that if the desired deployment is null and the actual
deployment is not null, the process performs deployment deletion.

Testing done:
Integration tests.

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com

---------

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com
Co-authored-by: github-actions <>