vdk-core: test ingestion with multiple threads
The IIngester methods (send_xxx_for_ingestion) should allow being called
from multiple threads. But it's never explicitly tested. This test is
far from perfect. Testing something is thread safe is not very easy. But
it's good to have it because A) it's indication of a contract for future
developers B) it's some verification still.