vdk-coverity: Adding Coverity Scan   (#2753)
**Why?**
The coverity scan is a SAST scan which is required for every product or
service to conform to Vmware SDL procedures.

**What?**
This job constitutes of a script which downloads the coverity-analysis.
Untars the analysis file and then executes the build for various
projects to create intermediate folder within the vdk repo. It then
combines the intermediate folder and uploads it to the
https://scan.coverity.com/.

**How has this been tested?**
Coverity Scan Test:
The job was triggered to run from a rule based on commit on a private
branch and the link for successful pipeline run is
here:https://gitlab.com/vmware-analytics/versatile-data-kit/-/pipelines/1024643834.
Only change was the commit branch has been changed to point to main
instead of a specific branch before merging.

**What type of change are you making?**
The change is in the gitlab-ci.yaml file where a new job has been added
as part of the existing pipeline configuration. This will create a new
stage/job when it is triggered.

---------

Co-authored-by: Shanmathi Kailasam <skailasam@vmware.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>