vdk-core: data sources POC
This change implements POC for Data Sources API.

It is based on some of teh requirements and reserach in
https://github.com/vmware/versatile-data-kit/wiki/Ingest-Source-Research

See concepts page for explanation of data source related concepts

So what's implemented is
- Data Source APIs handling sources, streams and state
- Data Source connection management partialy
- Data Source Ingester that reads from data sources and writes to
existing IIngeser
- An example data source AutoGeneratedDataSource
- An example job in the function test suite

Most likely this would be moved to plugin vdk-data-sources. For now it
doesn't appear ther's need for this to be in vdk-core.