vdk-core: test ingestion with multiple threads (#2796)
The IIngester methods (send_xxx_for_ingestion) should allow being called
from multiple threads. But it's never explicitly tested. This test is
far from perfect. Testing something is thread safe is not very easy.

But it's good to have it because

A) it's an indication of this thread-safe contract for future developers
B) If there is some serious regression this should catch it even if it
doesn't catch some more subtle regressions.