vdk-core: test exception propagation to user code
Why?

As part of our run logs initiative, we stopped wrapping non-vdk
errors into vdk errors. This should result in errors coming from
libraries, templates, etc. being propagated to user code. Users
should then be able to handle those errors. This behaviour has
not been tested, however.

What?

Add a functional test that creates a template and runs it. The
templates tries to read an empty csv using pandas, which results
in a pandas-specific exception. The exception is then handled
in a data job that executes the template.

How was this tested?

Ran the test locally

What kind of change is this?

Feature/non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>